Cleaned out EventEmitterPureComponent
Removed a lot of unused code in this component, and moved the 'upstate'
code we want to get rid of to the only remaining place where it's used
